feat: relocate weather utility functions and update import paths
